Extract cloud-cost-pricing shared package for cloudrift-iac-detector
cloudrift-iac-detector needs a $/month estimate for the zombie AWS
resources it finds, and should reuse this repo's pricing engine rather
than duplicate prices.json. cloud-cost-domain isn't a thin vocabulary
layer though — it's 44 resource-specific waste entities/policies plus
a ResourceKind union closed over exactly those 44 kinds, none of it
portable to another repo's own resource catalog.

Moves the genuinely generic slice (PricingPort, AwsRegion, CostEstimate,
WastePolicy, TablePricingAdapter/StaticPriceTableAdapter/prices.json)
into a new libs/shared/cloud-cost-pricing package, mirroring the
shared-aws-infra-utils publishing pattern (ADR-0085). Re-export shims
at every original location keep all six existing consumers' imports
unchanged. WastePolicy's generic bound moved from WastedResource to a
new minimal TaggedResource interface to avoid a circular dependency.

See ADR-0098 for the full rationale, including why AwsPricingApiAdapter
(live pricing) was deliberately left out of the extracted package.
