+++ libs/shared/cloud-cost-pricing/src/aws-region.value-object.spec.ts
