+++ libs/shared/cloud-cost-pricing/src/prices.json
