+++ libs/shared/cloud-cost-pricing/src/static-price-table.adapter.ts
@@ -1,0 +1,40 @@
+// SPDX-License-Identifier: Apache-2.0
+import priceTable from './prices.json';
+import {
+  TablePricingAdapter,
+  type PriceTable,
+  type RegionPrices,
+} from './table-pricing.adapter';
+
+function isRegionPricesEntry(entry: [string, unknown]): entry is [string, RegionPrices] {
+  return typeof entry[1] === 'object' && entry[1] !== null;
+}
+
+/**
+ * The built-in price table, extracted from `prices.json` by discarding the
+ * metadata fields (`_comment`, `pricesAsOf`) and keeping only the per-region
+ * tables. Exported so it can be composed with other sources (live API, user
+ * overrides).
+ */
+// Cast, not narrowed: the JSON module's inferred type has fixed, heterogeneous
+// keys (`_comment`/`pricesAsOf` are strings, each region is an object) rather
+// than a string index signature, so `Object.entries` can't iterate it
+// generically without first widening it to a `Record`. The actual filtering
+// (which entries are real per-region tables) is a real type predicate below,
+// not folded into this cast.
+export const BUILTIN_PRICE_TABLE: PriceTable = Object.fromEntries(
+  Object.entries(priceTable as Record<string, unknown>).filter(isRegionPricesEntry),
+);
+
+/** Date (YYYY-MM) the built-in price list was last verified. */
+export const BUILTIN_PRICES_AS_OF: string = priceTable.pricesAsOf;
+
+/**
+ * Pricing adapter based on the static `prices.json` price list. It is the
+ * always-available fallback when no live or manual prices are configured.
+ */
+export class StaticPriceTableAdapter extends TablePricingAdapter {
+  constructor() {
+    super(BUILTIN_PRICE_TABLE, BUILTIN_PRICES_AS_OF);
+  }
+}
