+++ libs/shared/cloud-cost-pricing/src/table-pricing.adapter.spec.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-import { AwsRegion } from 'cloud-cost-domain';
+import { AwsRegion } from './aws-region.value-object';
 import {
   TablePricingAdapter,
   mergePriceTables,
